backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
anthropic==0.42.0
chromadb==0.5.23
sentence-transformers==3.3.1
beautifulsoup4==4.12.3
requests==2.32.3
python-dotenv==1.0.1
redis==5.2.1
rank-bm25==0.2.2
numpy==1.26.4
